+++ src/main/java/basketbandit/core/modules/core/commands/CommandModules.java
@@ -50,7 +50,7 @@ public class CommandModules extends Command {
                     .setColor(Color.RED)
                     .setAuthor("Hey " + e.getAuthor().getName() + ",",null,e.getAuthor().getAvatarUrl())
                     .setTitle("Below are the list of bot module!")
-                    .setDescription("Each module can be toggled on or off by using the " + Configuration.PREFIX + "module <name> command.")
+                    .setDescription("Each module can be toggled on or off by using the " + Configuration.GLOBAL_PREFIX + "module <name> command.")
                     .addField("Enabled Modules", enabled.toString().replace(",","\n").replaceAll("[\\[\\] ]", "").toLowerCase(), false)
                     .addField("Disabled Modules", disabled.toString().replace(",","\n").replaceAll("[\\[\\] ]", "").toLowerCase(), false)
                     .setFooter("Version: " + Configuration.VERSION, e.getGuild().getMemberById(Configuration.BOT_ID).getUser().getAvatarUrl());