+++ src/main/java/basketbandit/core/modules/logging/ModuleLogging.java
@@ -29,7 +29,7 @@ public class ModuleLogging extends Module {
                 log.get(0).sendMessage("```" + e.getAuthor().getName() + "#" + e.getAuthor().getDiscriminator() + " used command " + e.getMessage().getContentDisplay() + " in " + e.getMessage().getChannel().getName() + ".```").queue();
             }
         } catch(Exception ex) {
-            e.getTextChannel().sendMessage("This server has logging enabled but has no 'command-log' text channel, please add that channel or disable the module by typing the '" + Configuration.PREFIX + "module logging' command to stop this message.").queue();
+            e.getTextChannel().sendMessage("This server has logging enabled but has no 'command-log' text channel, please add that channel or disable the module by typing the '" + Configuration.GLOBAL_PREFIX + "module logging' command to stop this message.").queue();
         }
     }
 }